migrated backend config out of Program
